+++ test/integration/048_rpc_test/deps_models/main.sql
@@ -1,0 +1,3 @@
+{{ dbt_utils.log_info('blah') }}
+
+select 1 as id
